+++ examples/simple.py
@@ -3,13 +3,16 @@ import asyncio
 
 from asphalt.core import ContainerComponent, Context, run_application
 
+from asphalt.serialization.api import Serializer
+
 
 class ApplicationComponent(ContainerComponent):
     async def start(self, ctx: Context):
         self.add_component("serialization", backend="json")
         await super().start(ctx)
 
-        payload = ctx.json.serialize(
+        serializer = ctx.require_resource(Serializer)
+        payload = serializer.serialize(
             {"a": 1, "b": 5.03, "c": [1, 2, 3], "d": {"x": "nested"}}
         )
         print("JSON serialized dict:", payload.decode())